+++ src/main/java/info/eecc/weather/exception/WeatherException.java
@@ -1,11 +1,22 @@
 package info.eecc.weather.exception;
 
+import org.springframework.http.HttpStatus;
+
 public class WeatherException extends RuntimeException {
-    public WeatherException(String message) {
+
+    private final HttpStatus httpStatus;
+
+    public WeatherException(HttpStatus httpStatus, String message) {
         super(message);
+        this.httpStatus = httpStatus;
     }
 
-    public WeatherException(String message, Throwable cause) {
+    public WeatherException(HttpStatus httpStatus, String message, Throwable cause) {
         super(message, cause);
+        this.httpStatus = httpStatus;
+    }
+
+    public HttpStatus getHttpStatus() {
+        return httpStatus;
     }
 }